keep definition as hidden column and include in search filter, ref DEV-274
