+++ libs/lib-glossary/src/lib/LandingPage.tsx
@@ -6,18 +6,12 @@ export const LandingPage = async () => {
   const client = createBrowserClient();
   const terms = await getAllGlossaryTerms({ client });
 
-  // Clean up the definition field by removing HTML tags
-  const rows = terms.map((term) => ({
-    ...term,
-    definition: term.definition.replace(/<[^>]+>/g, ''),
-  }));
-
   // extract types and languages from the terms data
   const types = Array.from(
-    new Set(rows.map((term) => term.type.split(', ')).flat()),
+    new Set(terms.map((term) => term.type.split(', ')).flat()),
   );
   const languages = Array.from(
-    new Set(rows.flatMap((term) => term.language)),
+    new Set(terms.flatMap((term) => term.language)),
   ).filter(Boolean);
 
   return (
@@ -25,7 +19,7 @@ export const LandingPage = async () => {
       <div className="w-full max-w-[1466px]">
         <H2 className="text-navy-500">{'Glossaries'}</H2>
         <GlossariesLandingTable
-          terms={rows}
+          terms={terms}
           types={types}
           languages={languages}
         />
