+++ libs/lib-glossary/src/ssr.ts
@@ -1,2 +1,3 @@
 export * from './lib/DetailPage';
 export * from './lib/LandingPage';
+export * from './lib/types';
